feat: allow initial chunks be anywhere
