+++ lib/ModuleSourceTypesConstants.js
@@ -45,6 +45,11 @@ const JS_TYPES = new Set(["javascript"]);
 const JS_AND_CSS_URL_TYPES = new Set(["javascript", "css-url"]);
 
 /**
+ * @type {ReadonlySet<"javascript" | "css">}
+ */
+const JS_AND_CSS_TYPES = new Set(["javascript", "css"]);
+
+/**
  * @type {ReadonlySet<"css">}
  */
 const CSS_TYPES = new Set(["css"]);
@@ -85,6 +90,7 @@ const SHARED_INIT_TYPES = new Set(["share-init"]);
 
 module.exports.NO_TYPES = NO_TYPES;
 module.exports.JS_TYPES = JS_TYPES;
+module.exports.JS_AND_CSS_TYPES = JS_AND_CSS_TYPES;
 module.exports.JS_AND_CSS_URL_TYPES = JS_AND_CSS_URL_TYPES;
 module.exports.ASSET_TYPES = ASSET_TYPES;
 module.exports.ASSET_AND_JS_TYPES = ASSET_AND_JS_TYPES;
