+++ lib/css/CssGenerator.js
@@ -5,37 +5,58 @@
 
 "use strict";
 
-const { ReplaceSource } = require("webpack-sources");
+const { ReplaceSource, RawSource, ConcatSource } = require("webpack-sources");
+const { UsageState } = require("../ExportsInfo");
 const Generator = require("../Generator");
 const InitFragment = require("../InitFragment");
-const { CSS_TYPES } = require("../ModuleSourceTypesConstants");
+const { JS_TYPES, JS_AND_CSS_TYPES } = require("../ModuleSourceTypesConstants");
 const RuntimeGlobals = require("../RuntimeGlobals");
+const Template = require("../Template");
 
 /** @typedef {import("webpack-sources").Source} Source */
-/** @typedef {import("../../declarations/WebpackOptions").CssGeneratorExportsConvention} CssGeneratorExportsConvention */
-/** @typedef {import("../../declarations/WebpackOptions").CssGeneratorLocalIdentName} CssGeneratorLocalIdentName */
+/** @typedef {import("../../declarations/WebpackOptions").CssAutoGeneratorOptions} CssAutoGeneratorOptions */
+/** @typedef {import("../../declarations/WebpackOptions").CssGlobalGeneratorOptions} CssGlobalGeneratorOptions */
+/** @typedef {import("../../declarations/WebpackOptions").CssModuleGeneratorOptions} CssModuleGeneratorOptions */
 /** @typedef {import("../CodeGenerationResults")} CodeGenerationResults */
 /** @typedef {import("../Dependency")} Dependency */
 /** @typedef {import("../DependencyTemplate").CssDependencyTemplateContext} DependencyTemplateContext */
 /** @typedef {import("../DependencyTemplate").CssExportsData} CssExportsData */
 /** @typedef {import("../Generator").GenerateContext} GenerateContext */
 /** @typedef {import("../Generator").UpdateHashContext} UpdateHashContext */
+/** @typedef {import("../Module").ConcatenationBailoutReasonContext} ConcatenationBailoutReasonContext */
 /** @typedef {import("../Module").SourceTypes} SourceTypes */
 /** @typedef {import("../NormalModule")} NormalModule */
 /** @typedef {import("../util/Hash")} Hash */
 
 class CssGenerator extends Generator {
 	/**
-	 * @param {CssGeneratorExportsConvention} convention the convention of the exports name
-	 * @param {CssGeneratorLocalIdentName} localIdentName css export local ident name
-	 * @param {boolean} esModule whether to use ES modules syntax
+	 * @param {CssAutoGeneratorOptions | CssGlobalGeneratorOptions | CssModuleGeneratorOptions} options options
 	 */
-	constructor(convention, localIdentName, esModule) {
+	constructor(options) {
 		super();
-		this.convention = convention;
-		this.localIdentName = localIdentName;
-		/** @type {boolean} */
-		this.esModule = esModule;
+		this.convention = options.exportsConvention;
+		this.localIdentName = options.localIdentName;
+		this.exportsOnly = options.exportsOnly;
+		this.esModule = options.esModule;
+	}
+
+	/**
+	 * @param {NormalModule} module module for which the bailout reason should be determined
+	 * @param {ConcatenationBailoutReasonContext} context context
+	 * @returns {string | undefined} reason why this module can't be concatenated, undefined when it can be concatenated
+	 */
+	getConcatenationBailoutReason(module, context) {
+		if (!this.esModule) {
+			return "Module is not an ECMAScript module";
+		}
+		// TODO webpack 6: remove /\[moduleid\]/.test
+		if (
+			/\[id\]/.test(this.localIdentName) ||
+			/\[moduleid\]/.test(this.localIdentName)
+		) {
+			return "The localIdentName includes moduleId ([id] or [moduleid])";
+		}
+		return undefined;
 	}
 
 	/**
@@ -44,8 +65,11 @@ class CssGenerator extends Generator {
 	 * @returns {Source | null} generated code
 	 */
 	generate(module, generateContext) {
-		const originalSource = /** @type {Source} */ (module.originalSource());
-		const source = new ReplaceSource(originalSource);
+		const source =
+			generateContext.type === "javascript"
+				? new ReplaceSource(new RawSource(""))
+				: new ReplaceSource(/** @type {Source} */ (module.originalSource()));
+
 		/** @type {InitFragment<GenerateContext>[]} */
 		const initFragments = [];
 		/** @type {CssExportsData} */
@@ -53,8 +77,6 @@ class CssGenerator extends Generator {
 			esModule: this.esModule,
 			exports: new Map()
 		};
-
-		generateContext.runtimeRequirements.add(RuntimeGlobals.hasCssModules);
 
 		/** @type {InitFragment<GenerateContext>[] | undefined} */
 		let chunkInitFragments;
@@ -105,21 +127,79 @@ class CssGenerator extends Generator {
 
 			template.apply(dependency, source, templateContext);
 		};
+
 		for (const dependency of module.dependencies) {
 			handleDependency(dependency);
 		}
-		if (module.presentationalDependencies !== undefined) {
-			for (const dependency of module.presentationalDependencies) {
-				handleDependency(dependency);
+
+		switch (generateContext.type) {
+			case "javascript": {
+				generateContext.runtimeRequirements.add(RuntimeGlobals.module);
+
+				if (generateContext.concatenationScope) {
+					const source = new ConcatSource();
+					const usedIdentifiers = new Set();
+					for (const [name, v] of cssExportsData.exports) {
+						let identifier = Template.toIdentifier(name);
+						const i = 0;
+						while (usedIdentifiers.has(identifier)) {
+							identifier = Template.toIdentifier(name + i);
+						}
+						usedIdentifiers.add(identifier);
+						generateContext.concatenationScope.registerExport(name, identifier);
+						source.add(
+							`${
+								generateContext.runtimeTemplate.supportsConst()
+									? "const"
+									: "var"
+							} ${identifier} = ${JSON.stringify(v)};\n`
+						);
+					}
+					return source;
+				}
+
+				const needNsObj =
+					this.esModule &&
+					generateContext.moduleGraph
+						.getExportsInfo(module)
+						.otherExportsInfo.getUsed(generateContext.runtime) !==
+						UsageState.Unused;
+
+				if (needNsObj) {
+					generateContext.runtimeRequirements.add(
+						RuntimeGlobals.makeNamespaceObject
+					);
+				}
+
+				const exports = [];
+
+				for (const [name, v] of cssExportsData.exports) {
+					exports.push(`\t${JSON.stringify(name)}: ${JSON.stringify(v)}`);
+				}
+
+				return new RawSource(
+					`${needNsObj ? `${RuntimeGlobals.makeNamespaceObject}(` : ""}${
+						module.moduleArgument
+					}.exports = {\n${exports.join(",\n")}\n}${needNsObj ? ")" : ""};`
+				);
+			}
+			case "css": {
+				if (module.presentationalDependencies !== undefined) {
+					for (const dependency of module.presentationalDependencies) {
+						handleDependency(dependency);
+					}
+				}
+
+				generateContext.runtimeRequirements.add(RuntimeGlobals.hasCssModules);
+
+				const data =
+					/** @type {NonNullable<GenerateContext["getData"]>} */
+					(generateContext.getData)();
+				data.set("css-exports", cssExportsData);
+
+				return InitFragment.addToSource(source, initFragments, generateContext);
 			}
 		}
-
-		const data =
-			/** @type {NonNullable<GenerateContext["getData"]>} */
-			(generateContext.getData)();
-		data.set("css-exports", cssExportsData);
-
-		return InitFragment.addToSource(source, initFragments, generateContext);
 	}
 
 	/**
@@ -127,7 +207,7 @@ class CssGenerator extends Generator {
 	 * @returns {SourceTypes} available types (do not mutate)
 	 */
 	getTypes(module) {
-		return CSS_TYPES;
+		return this.exportsOnly ? JS_TYPES : JS_AND_CSS_TYPES;
 	}
 
 	/**
@@ -136,13 +216,20 @@ class CssGenerator extends Generator {
 	 * @returns {number} estimate size of the module
 	 */
 	getSize(module, type) {
-		const originalSource = module.originalSource();
+		switch (type) {
+			case "javascript": {
+				return 42;
+			}
+			case "css": {
+				const originalSource = module.originalSource();
 
-		if (!originalSource) {
-			return 0;
+				if (!originalSource) {
+					return 0;
+				}
+
+				return originalSource.size();
+			}
 		}
-
-		return originalSource.size();
 	}
 
 	/**
