+++ lib/css/CssLoadingRuntimeModule.js
@@ -98,17 +98,14 @@ class CssLoadingRuntimeModule extends RuntimeModule {
 			RuntimeGlobals.hmrDownloadUpdateHandlers
 		);
 		/** @type {Set<number | string | null>} */
-		const initialChunkIdsWithCss = new Set();
-		/** @type {Set<number | string | null>} */
-		const initialChunkIdsWithoutCss = new Set();
+		const initialChunkIds = new Set();
 		for (const c of /** @type {Chunk} */ (chunk).getAllInitialChunks()) {
-			(chunkHasCss(c, chunkGraph)
-				? initialChunkIdsWithCss
-				: initialChunkIdsWithoutCss
-			).add(c.id);
+			if (chunkHasCss(c, chunkGraph)) {
+				initialChunkIds.add(c.id);
+			}
 		}
 
-		if (!withLoading && !withHmr && initialChunkIdsWithCss.size === 0) {
+		if (!withLoading && !withHmr) {
 			return null;
 		}
 
@@ -185,10 +182,13 @@ class CssLoadingRuntimeModule extends RuntimeModule {
 			"// [resolve, reject, Promise] = chunk loading, 0 = chunk loaded",
 			`var installedChunks = ${
 				stateExpression ? `${stateExpression} = ${stateExpression} || ` : ""
-			}{${Array.from(
-				initialChunkIdsWithoutCss,
-				id => `${JSON.stringify(id)}:0`
-			).join(",")}};`,
+			}{`,
+			Template.indent(
+				Array.from(initialChunkIds, id => `${JSON.stringify(id)}: 0`).join(
+					",\n"
+				)
+			),
+			"};",
 			"",
 			uniqueName
 				? `var uniqueName = ${JSON.stringify(
@@ -203,7 +203,6 @@ class CssLoadingRuntimeModule extends RuntimeModule {
 					}name = ${name}, i, cc = 1;`,
 					"try {",
 					Template.indent([
-						"if(!link) link = loadStylesheet(chunkId);",
 						// `link.sheet.rules` for legacy browsers
 						"var cssRules = link.sheet.cssRules || link.sheet.rules;",
 						"var j = cssRules.length - 1;",
@@ -324,21 +323,6 @@ class CssLoadingRuntimeModule extends RuntimeModule {
 					"return link;"
 				]
 			)};`,
-			initialChunkIdsWithCss.size > 2
-				? `${JSON.stringify(
-						Array.from(initialChunkIdsWithCss)
-					)}.forEach(loadCssChunkData.bind(null, ${
-						RuntimeGlobals.moduleFactories
-					}, 0));`
-				: initialChunkIdsWithCss.size > 0
-					? `${Array.from(
-							initialChunkIdsWithCss,
-							id =>
-								`loadCssChunkData(${
-									RuntimeGlobals.moduleFactories
-								}, 0, ${JSON.stringify(id)});`
-						).join("")}`
-					: "// no initial css",
 			"",
 			withLoading
 				? Template.asString([
