+++ lib/css/CssModulesPlugin.js
@@ -39,7 +39,6 @@ const createHash = require("../util/createHash");
 const { getUndoPath } = require("../util/identifier");
 const memoize = require("../util/memoize");
 const nonNumericOnlyHash = require("../util/nonNumericOnlyHash");
-const CssExportsGenerator = require("./CssExportsGenerator");
 const CssGenerator = require("./CssGenerator");
 const CssParser = require("./CssParser");
 
@@ -344,17 +343,7 @@ class CssModulesPlugin {
 						.tap(PLUGIN_NAME, generatorOptions => {
 							validateGeneratorOptions[type](generatorOptions);
 
-							return generatorOptions.exportsOnly
-								? new CssExportsGenerator(
-										generatorOptions.exportsConvention,
-										generatorOptions.localIdentName,
-										generatorOptions.esModule
-									)
-								: new CssGenerator(
-										generatorOptions.exportsConvention,
-										generatorOptions.localIdentName,
-										generatorOptions.esModule
-									);
+							return new CssGenerator(generatorOptions);
 						});
 					normalModuleFactory.hooks.createModuleClass
 						.for(type)
